# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WikiLLM API
  description: |-
    HTTP API for [wiki-llm](https://gist.github.com/karpathy/442a6bf555914893e9891c11519de94f)
    folders. Designed to run alongside Obsidian, git, LiveSync, and other sync tools while
    serving multiple users and agents from the same markdown wiki.

    ## Project links

    - [GitHub repository](https://github.com/LukasParke/wikillm-api)
    - [Docker image](https://github.com/LukasParke/wikillm-api/pkgs/container/wikillm-api)
    - [Deployment guide](https://github.com/LukasParke/wikillm-api#deployment)
    - [Architecture docs](https://github.com/LukasParke/wikillm-api/blob/main/docs/ARCHITECTURE.md)

    ## Features

    - Self-hosted knowledge base service for humans and agents
    - Filesystem remains the source of truth; the index is a rebuildable cache
    - Atomic writes with optimistic concurrency control
    - Multi-source API key attribution with OKF actor convention
    - Hybrid retrieval: full-text + vector + recency decay, RRF fusion (K=60), optional rerank
    - LLM answering with citations (planner → executor → synthesis)
    - SSE and WebSocket live change feeds
    - Batch ingestion of sources and wiki pages
    - Connector framework (git, web, GitHub)
    - MCP server exposing LLM-free retrieval tools
    - Projects scoping with per-key RBAC (read/write/admin)
    - Google OKF v0.2 conformance: validation, trust tiers, provenance stamping
  version: 0.1.0
  contact:
    name: WikiLLM API
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key passed as a Bearer token. Keys are configured via the API_KEYS environment variable.
  schemas:
...
    Conflict:
      type: object
      required:
        - hash
        - content
      properties:
        hash:
          type: string
        content:
          type: string
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: conflict
        message:
          type: string
        current:
          $ref: "#/components/schemas/Conflict"
        existingHash:
          type: string
...
    WebhookInput:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          default: ["change"]
          items:
            type: string
            enum:
              - change
        prefixes:
          type: array
          default: ["*"]
          description: Path prefixes that trigger delivery
          items:
            type: string
        enabled:
          type: boolean
          default: true
    Webhook:
      type: object
      required:
        - id
        - url
        - events
        - prefixes
        - enabled
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
              - change
        prefixes:
          type: array
          items:
            type: string
        enabled:
          type: boolean
        last_status:
          type: integer
          nullable: true
          description: HTTP status of the last delivery attempt
        last_attempt_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    WebhookList:
      type: object
      required:
        - webhooks
      properties:
        webhooks:
          type: array
          items:
            $ref: "#/components/schemas/Webhook"
paths:
...
  /v1/webhooks:
    get:
      tags:
        - Webhooks
      summary: List outbound webhooks
      description: Requires admin role.
      responses:
        "200":
          description: Registered webhooks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookList"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Webhooks
      summary: Register an outbound webhook
      description: |
        Deliveries are HMAC-SHA256 signed (X-WikiLLM-Signature header) using the
        webhook_secret runtime setting and retried at 250ms, 1s, and 4s. Each
        webhook tracks last_status and last_attempt_at.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookInput"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/webhooks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Webhook id

    delete:
      tags:
        - Webhooks
      summary: Delete a webhook
      responses:
        "200":
          description: Webhook deleted
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Unknown webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...